feat(network): extract get_connected_stream_futures to test utils (#1167)
* test(network): move get_connected_stream_futures to other file

* test(network): remove futures from get_connected_stream_futures
