+++ crates/papyrus_network/src/lib.rs
@@ -4,6 +4,8 @@
 /// [`Starknet p2p specs`]: https://github.com/starknet-io/starknet-p2p-specs/
 pub mod messages;
 pub mod streamed_data_protocol;
+#[cfg(test)]
+mod test_utils;
 
 use starknet_api::block::{BlockHash, BlockNumber};
